Refactor TP3 data handling and messaging: update INode interface to use Id instead of Qid; enhance TP3ReadRequest and TP3ReadResponse for improved data streaming; implement DirectoryStreamData for directory handling; streamline read operations across various components; add ReadResponseDataStream for efficient chunked reading; improve error handling and logging throughout the system.

+++ src/TP3.CLI/MessageHelper.cs
@@ -42,8 +42,8 @@ public class MessageHelper
             return new TP3ReadRequest
             {
                 Tag = tag,
-                Offset = offset,
-                MaxBytes = maxBytes,
+                Offset = (ulong)offset,
+                MaxBytes = (uint)maxBytes,
             };
         }
